feat: remove trailing zeroes from Integers

+++ src/main/text_math/decimal.hpp
@@ -1,12 +1,11 @@
 #ifndef DECIMAL_HPP
 #define DECIMAL_HPP
 #include "number.hpp"
-#include "integer.hpp"
 
 namespace TextMath {
     class Decimal final : public Number {
     private:
-        Integer precision = 2;
+        size_t precision = 2;
 
     public:
         Decimal() noexcept;
